Snake: Move geometry types into Geometry.h and add Direction enum
